Move Aggregate GenServer to client side via `use/1` (...)
The `GenServer`  behaviour is now  injected to user  defined aggregates,
one only  has to `use Commanded.Aggregates.Aggregate`  (no options yet),
and the callbacks will be defined as well.

The  motivation behind  this change  was to  make the  aggregate process
configurable. For example, what if at point `GenServer` is replaced with
`GenStatem`,  adding  the  capability  to  explicitly  define  aggregate
states, but  would provide  backwards compatibility  to use  the current
semantics?